refactor(diagnostics): move warning machinery into warning module
